notifications: reorganize within a dedicated controller folder
