+++ server/controllers/notifications/lib/delete_notifications.js
@@ -1,0 +1,10 @@
+const __ = require('config').universalPath
+const _ = __.require('builders', 'utils')
+const assert_ = __.require('utils', 'assert_types')
+const notifications_ = require('./notifications')
+
+module.exports = (label, subjectId) => {
+  assert_.strings([ label, subjectId ])
+  _.log(`deleting ${label} notifications`)
+  return notifications_.deleteAllBySubjectId(subjectId)
+}
