+++ server/controllers/notifications/lib/user_made_admin.js
@@ -1,0 +1,11 @@
+const __ = require('config').universalPath
+const assert_ = __.require('utils', 'assert_types')
+const notifications_ = require('./notifications')
+
+module.exports = (groupId, actorAdminId, newAdminId) => {
+  assert_.strings([ groupId, actorAdminId, newAdminId ])
+  return notifications_.add(newAdminId, 'userMadeAdmin', {
+    group: groupId,
+    user: actorAdminId
+  })
+}
